cleanup: move parses_instruction to insts module
