--- src/insts.rs
@@ -1,5 +1,0 @@
-//! `insts` is a module that contains the definitions of WebAssembly instructions.
-
-pub mod control;
-pub mod memory;
-pub mod numeric;
